+++ src/Admin/AdminPages.php
@@ -561,7 +561,10 @@ class AdminPages {
         
         $settings['auto_log_cleanup'] = isset( $_POST['auto_log_cleanup'] );
         $settings['debug_mode'] = isset( $_POST['debug_mode'] );
-        
+
+        // Data management settings
+        $settings['skip_zero_change_scans'] = isset( $_POST['skip_zero_change_scans'] );
+
         // Uninstall settings
         $settings['delete_data_on_uninstall'] = isset( $_POST['delete_data_on_uninstall'] );